backend/pyproject.toml
[project]
name = "smart-keep-backend"
version = "0.1.0"
description = "FastAPI backend for Smart Keep."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.29,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "email-validator>=2.1,<3.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "bcrypt>=4.0.1,<4.1",
    "google-generativeai>=0.7,<1.0",
]

[tool.uv]
package = false
